// types/match.ts
export type Participant = {
    allInPings: number;
    assistMePings: number;
    assists: number;
    baronKills: number;
    basicPings: number;
    bountyLevel: number;
    challenges: Challenges;
    champExperience: number;
    champLevel: number;
    championId: number;
    championName: string;
    championTransform: number;
    commandPings: number;
    consumablesPurchased: number;
    damageDealtToBuildings: number;
    damageDealtToObjectives: number;
    damageDealtToTurrets: number;
    damageSelfMitigated: number;
    dangerPings: number;
    deaths: number;
    detectorWardsPlaced: number;
    doubleKills: number;
    dragonKills: number;
    eligibleForProgression: boolean;
    enemyMissingPings: number;
    enemyVisionPings: number;
    firstBloodAssist: boolean;
    firstBloodKill: boolean;
    firstTowerAssist: boolean;
    firstTowerKill: boolean;
    gameEndedInEarlySurrender: boolean;
    gameEndedInSurrender: boolean;
    getBackPings: number;
    goldEarned: number;
    goldSpent: number;
    holdPings: number;
    individualPosition: string;
    inhibitorKills: number;
    inhibitorTakedowns: number;
    inhibitorsLost: number;
    item0: number;
    item1: number;
    item2: number;
    item3: number;
    item4: number;
    item5: number;
    item6: number;
    itemsPurchased: number;
    killingSprees: number;
    kills: number;
    lane: string;
    largestCriticalStrike: number;
    largestKillingSpree: number;
    largestMultiKill: number;
    longestTimeSpentLiving: number;
    magicDamageDealt: number;
    magicDamageDealtToChampions: number;
    magicDamageTaken: number;
    missions: Missions;
    needVisionPings: number;
    neutralMinionsKilled: number;
    nexusKills: number;
    nexusLost: number;
    nexusTakedowns: number;
    objectivesStolen: number;
    objectivesStolenAssists: number;
    onMyWayPings: number;
    participantId: number;
    pentaKills: number;
    perks: Perks;
    physicalDamageDealt: number;
    physicalDamageDealtToChampions: number;
    physicalDamageTaken: number;
    placement: number;
    playerAugment1: number;
    playerAugment2: number;
    playerAugment3: number;
    playerAugment4: number;
    playerAugment5: number;
    playerAugment6: number;
    playerSubteamId: number;
    profileIcon: number;
    pushPings: number;
    puuid: string;
    quadraKills: number;
    retreatPings: number;
    riotIdGameName: string;
    riotIdTagline: string;
    role: string;
    sightWardsBoughtInGame: number;
    spell1Casts: number;
    spell2Casts: number;
    spell3Casts: number;
    spell4Casts: number;
    subteamPlacement: number;
    summoner1Casts: number;
    summoner1Id: number;
    summoner2Casts: number;
    summoner2Id: number;
    summonerId: string;
    summonerLevel: number;
    summonerName: string;
    teamEarlySurrendered: boolean;
    teamId: number;
    teamPosition: string;
    timeCCingOthers: number;
    timePlayed: number;
    totalAllyJungleMinionsKilled: number;
    totalDamageDealt: number;
    totalDamageDealtToChampions: number;
    totalDamageShieldedOnTeammates: number;
    totalDamageTaken: number;
    totalEnemyJungleMinionsKilled: number;
    totalHeal: number;
    totalHealsOnTeammates: number;
    totalMinionsKilled: number;
    totalTimeCCDealt: number;
    totalTimeSpentDead: number;
    totalUnitsHealed: number;
    tripleKills: number;
    trueDamageDealt: number;
    trueDamageDealtToChampions: number;
    trueDamageTaken: number;
    turretKills: number;
    turretTakedowns: number;
    turretsLost: number;
    unrealKills: number;
    visionClearedPings: number;
    visionScore: number;
    visionWardsBoughtInGame: number;
    wardsKilled: number;
    wardsPlaced: number;
    win: boolean;
};

export type Perks = {
    statPerks: {
        defense: number;
        flex: number;
        offense: number;
    };
    styles: {
        description: string;
        selections: {
            perk: number;
            var1: number;
            var2: number;
            var3: number;
        }[];
        style: number;
    }[];
};
// ...
export type Challenges = {
    '12AssistStreakCount'?: number;
    HealFromMapSources?: number;
    InfernalScalePickup?: number;
    SWARM_DefeatAatrox?: number;
    SWARM_DefeatBriar?: number;
    SWARM_DefeatMiniBosses?: number;
    SWARM_EvolveWeapon?: number;
    SWARM_Have3Passives?: number;
    SWARM_KillEnemy?: number;
    SWARM_PickupGold?: number;
    SWARM_ReachLevel50?: number;
    SWARM_Survive15Min?: number;
    SWARM_WinWith5EvolvedWeapons?: number;
    abilityUses?: number;
    acesBefore15Minutes?: number;
    alliedJungleMonsterKills?: number;
    baronBuffGoldAdvantageOverThreshold?: number;
    baronTakedowns?: number;
    blastConeOppositeOpponentCount?: number;
    bountyGold?: number;
    buffsStolen?: number;
    completeSupportQuestInTime?: number;
    controlWardTimeCoverageInRiverOrEnemyHalf?: number;
    controlWardsPlaced?: number;
    damagePerMinute?: number;
    damageTakenOnTeamPercentage?: number;
    dancedWithRiftHerald?: number;
    deathsByEnemyChamps?: number;
    dodgeSkillShotsSmallWindow?: number;
    doubleAces?: number;
    dragonTakedowns?: number;
    earliestBaron?: number;
    earliestDragonTakedown?: number;
    earlyLaningPhaseGoldExpAdvantage?: number;
    effectiveHealAndShielding?: number;
    elderDragonKillsWithOpposingSoul?: number;
    elderDragonMultikills?: number;
    enemyChampionImmobilizations?: number;
    enemyJungleMonsterKills?: number;
    epicMonsterKillsNearEnemyJungler?: number;
    epicMonsterKillsWithin30SecondsOfSpawn?: number;
    epicMonsterSteals?: number;
    epicMonsterStolenWithoutSmite?: number;
    firstTurretKilled?: number;
    firstTurretKilledTime?: number;
    fistBumpParticipation?: number;
    flawlessAces?: number;
    fullTeamTakedown?: number;
    gameLength?: number;
    getTakedownsInAllLanesEarlyJungleAsLaner?: number;
    goldPerMinute?: number;
    hadOpenNexus?: number;
    immobilizeAndKillWithAlly?: number;
    initialBuffCount?: number;
    initialCrabCount?: number;
    jungleCsBefore10Minutes?: number;
    junglerTakedownsNearDamagedEpicMonster?: number;
    kTurretsDestroyedBeforePlatesFall?: number;
    kda?: number;
    killAfterHiddenWithAlly?: number;
    killParticipation?: number;
    killedChampTookFullTeamDamageSurvived?: number;
    killingSprees?: number;
    killsNearEnemyTurret?: number;
    killsOnOtherLanesEarlyJungleAsLaner?: number;
    killsOnRecentlyHealedByAramPack?: number;
    killsUnderOwnTurret?: number;
    killsWithHelpFromEpicMonster?: number;
    knockEnemyIntoTeamAndKill?: number;
    landSkillShotsEarlyGame?: number;
    laneMinionsFirst10Minutes?: number;
    laningPhaseGoldExpAdvantage?: number;
    legendaryCount?: number;
    legendaryItemUsed?: number[];
    lostAnInhibitor?: number;
    maxCsAdvantageOnLaneOpponent?: number;
    maxKillDeficit?: number;
    maxLevelLeadLaneOpponent?: number;
    mejaisFullStackInTime?: number;
    moreEnemyJungleThanOpponent?: number;
    multiKillOneSpell?: number;
    multiTurretRiftHeraldCount?: number;
    multikills?: number;
    multikillsAfterAggressiveFlash?: number;
    outerTurretExecutesBefore10Minutes?: number;
    outnumberedKills?: number;
    outnumberedNexusKill?: number;
    perfectDragonSoulsTaken?: number;
    perfectGame?: number;
    pickKillWithAlly?: number;
    playedChampSelectPosition?: number;
    poroExplosions?: number;
    quickCleanse?: number;
    quickFirstTurret?: number;
    quickSoloKills?: number;
    riftHeraldTakedowns?: number;
    saveAllyFromDeath?: number;
    scuttleCrabKills?: number;
    skillshotsDodged?: number;
    skillshotsHit?: number;
    snowballsHit?: number;
    soloBaronKills?: number;
    soloKills?: number;
    soloTurretsLategame?: number;
    stealthWardsPlaced?: number;
    survivedSingleDigitHpCount?: number;
    survivedThreeImmobilizesInFight?: number;
    takedownOnFirstTurret?: number;
    takedowns?: number;
    takedownsAfterGainingLevelAdvantage?: number;
    takedownsBeforeJungleMinionSpawn?: number;
    takedownsFirstXMinutes?: number;
    takedownsInAlcove?: number;
    takedownsInEnemyFountain?: number;
    teamBaronKills?: number;
    teamDamagePercentage?: number;
    teamElderDragonKills?: number;
    teamRiftHeraldKills?: number;
    tookLargeDamageSurvived?: number;
    turretPlatesTaken?: number;
    turretTakedowns?: number;
    turretsTakenWithRiftHerald?: number;
    twentyMinionsIn3SecondsCount?: number;
    twoWardsOneSweeperCount?: number;
    unseenRecalls?: number;
    visionScoreAdvantageLaneOpponent?: number;
    visionScorePerMinute?: number;
    voidMonsterKill?: number;
    wardTakedowns?: number;
    wardTakedownsBefore20M?: number;
    wardsGuarded?: number;
    junglerKillsEarlyJungle?: number;
    killsOnLanersEarlyJungleAsJungler?: number;
    shortestTimeToAceFromFirstTakedown?: number;
    highestWardKills?: number;
    highestChampionDamage?: number;
    fasterSupportQuestCompletion?: number;
    highestCrowdControlScore?: number;
};

export type Missions = {
    playerScore0: number;
    playerScore1: number;
    playerScore2: number;
    playerScore3: number;
    playerScore4: number;
    playerScore5: number;
    playerScore6: number;
    playerScore7: number;
    playerScore8: number;
    playerScore9: number;
    playerScore10: number;
    playerScore11: number;
};